Develop

+++ src/main/java/com/credential/cubrism/server/posts/dto/GPTRequest.java
@@ -12,21 +12,20 @@ import java.util.List;
 public class GPTRequest {
     private String model;
     private List<Message> messages;
-    private int temperature;
+    private double temperature;
     private int maxTokens;
     private int topP;
     private int frequencyPenalty;
     private int presencePenalty;
 
-    public GPTRequest(String model, String prompt, int temperature, int maxTokens, int topP, int frequencyPenalty, int presencePenalty) {
+    public GPTRequest(String model, String prompt, double temperature, int maxTokens, int topP, int frequencyPenalty, int presencePenalty) {
         this.model = model;
         this.messages = new ArrayList<>();
-        this.messages.add(new Message("user",prompt));
+        this.messages.add(new Message("user", prompt));
         this.temperature = temperature;
         this.maxTokens = maxTokens;
-        this.topP=topP;
-        this.frequencyPenalty=frequencyPenalty;
+        this.topP = topP;
+        this.frequencyPenalty = frequencyPenalty;
         this.presencePenalty = presencePenalty;
-
     }
 }
